Dataset has one license, either open or closed
